backend/src/requirements.txt
asgiref==3.10.0
channels==4.3.1
channels_redis==4.3.0
Django==5.2.7
djangorestframework==3.16.1
msgpack==1.1.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
pillow==12.0.0
redis==7.0.0
sqlparse==0.5.3
tzdata==2025.2
